spark-stand-alone-log-aggregaor/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
pydantic==1.8.2
python-multipart==0.0.5
passlib==1.7.4
async-exit-stack==1.0.1
async-generator==1.10
aiofiles==0.7.0
itsdangerous==2.0.1
requests==2.25.1
PyYAML==5.4.1
parsel~=1.6.0
prometheus-client==0.11.0
